Added Code Fences
The code fences allow the class to be folded up in the IDE and are
easier to find when skimming through a large chunk of code than the
class definition.

Not sure if Electric Imp wants this boilerplate on in their classes so I
am making a seperate commit so that it can be removed later.